split and splitabstain support
