+++ src/lib/utils.ts
@@ -4,7 +4,12 @@ import networks from '@/assets/networks.json'
 
 import { twMerge } from 'tailwind-merge'
 import type { NetworkType, RouterType, Vote } from './types'
-import { NetworksFromConfig } from '@/contexts/NetworkContext'
+import { ApiType, NetworksFromConfig } from '@/contexts/NetworkContext'
+import { MultiAddress } from '@polkadot-api/descriptors'
+import { InjectedPolkadotAccount } from 'polkadot-api/pjs-signer'
+import { DEFAULT_TIME, lockPeriod, ONE_DAY, THRESHOLD } from './constants'
+import { bnMin } from './bnMin'
+import { VoteLock } from '@/contexts/LocksContext'
 
 export const cn = (...inputs: ClassValue[]) => {
   return twMerge(clsx(inputs))
@@ -29,3 +34,73 @@ export const getVoteFromNumber = (input: number): Vote => ({
 
 export const getNumberFromVote = ({ aye, conviction }: Vote): number =>
   +aye * 0b1000_0000 + conviction
+
+export const getUnlockUnvoteTx = (
+  locks: VoteLock[],
+  api: ApiType,
+  account: InjectedPolkadotAccount,
+) => {
+  const tracks = new Set(locks.map((lock) => lock.trackId))
+
+  const unVoteTxs = locks.map((lock) => {
+    return api.tx.ConvictionVoting.remove_vote({
+      index: lock.refId,
+      class: lock.trackId,
+    }).decodedCall
+  })
+  const unlockTxs = Array.from(tracks).map((trackId) => {
+    return api.tx.ConvictionVoting.unlock({
+      class: trackId,
+      target: MultiAddress.Id(account.address),
+    }).decodedCall
+  })
+
+  return { unVoteTxs, unlockTxs }
+}
+
+export const indexToConviction = (index: number) => {
+  return Object.keys(lockPeriod)[index]
+}
+
+const convictionList = Object.keys(lockPeriod)
+
+export const getExpectedBlockTimeMs = async (api: ApiType): Promise<bigint> => {
+  const expectedBlockTime = await api.constants.Babe.ExpectedBlockTime()
+  if (expectedBlockTime) {
+    return bnMin(ONE_DAY, expectedBlockTime)
+  }
+
+  const thresholdCheck =
+    (await api.constants.Timestamp.MinimumPeriod()) > THRESHOLD
+
+  if (thresholdCheck) {
+    return bnMin(ONE_DAY, (await api.constants.Timestamp.MinimumPeriod()) * 2n)
+  }
+
+  return bnMin(ONE_DAY, DEFAULT_TIME)
+}
+
+export const getLockTimes = async (api: ApiType) => {
+  const voteLockingPeriodBlocks =
+    await api.constants.ConvictionVoting.VoteLockingPeriod()
+
+  const expectedBlockTimeMs = await getExpectedBlockTimeMs(api)
+
+  const requests = convictionList.map((conviction) => {
+    const lockTimeMs =
+      expectedBlockTimeMs *
+      BigInt(voteLockingPeriodBlocks) *
+      BigInt(lockPeriod[conviction])
+
+    return [conviction, lockTimeMs] as const
+  })
+
+  return requests.reduce(
+    (acc, [conviction, lockPeriod]) => {
+      acc[conviction] = lockPeriod
+
+      return acc
+    },
+    {} as Record<string, bigint>,
+  )
+}
