+++ src/pages/Delegate/index.tsx
@@ -2,7 +2,7 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { useDelegates } from '@/contexts/DelegatesContext'
 import { useNetwork } from '@/contexts/NetworkContext'
-import { getLockTimes } from '@/lib/locks'
+import { getLockTimes } from '@/lib/utils'
 import { VotingConviction } from '@polkadot-api/descriptors'
 import { SetStateAction, useEffect, useState } from 'react'
 import { convertMiliseconds } from '@/lib/convertMiliseconds'
